+++ enterprise/frontend/src/metabase-enterprise/advanced_permissions/components/ImpersonationModal/ImpersonationModal.unit.spec.tsx
@@ -113,7 +113,7 @@ describe("impersonation modal", () => {
 
     expect(
       await screen.findByRole("link", { name: /edit settings/i }),
-    ).toHaveAttribute("href", "/admin/databases/1");
+    ).toHaveAttribute("href", "/admin/databases/1/edit");
   });
 
   it("should refer to 'users' instead of 'roles' for redshift impersonation", async () => {
