+++ frontend/src/metabase/admin/databases/components/DatabaseDangerZoneSection/index.tsx
@@ -1,0 +1,1 @@
+export * from "./DatabaseDangerZoneSection";
